Add quick actions section to new asset page
