+++ apps/ledger-live-mobile/src/components/FabActions/hooks/useAssetActions.tsx
@@ -1,0 +1,194 @@
+import React, { useMemo } from "react";
+import { AccountLike, Account } from "@ledgerhq/types-live";
+import { useSelector } from "react-redux";
+import { useTranslation } from "react-i18next";
+import { Icons } from "@ledgerhq/native-ui";
+import { NavigatorName, ScreenName } from "../../../const";
+// eslint-disable-next-line import/named
+import {
+  readOnlyModeEnabledSelector,
+  swapSelectableCurrenciesSelector,
+} from "../../../reducers/settings";
+import { useRampCatalog } from "@ledgerhq/live-common/platform/providers/RampCatalogProvider/index";
+import { filterRampCatalogEntries } from "@ledgerhq/live-common/platform/providers/RampCatalogProvider/helpers";
+import { ActionButton } from "../../../components/FabActions";
+import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
+
+type useAssetActionsProps = {
+  currency?: CryptoCurrency;
+  accounts?: AccountLike[];
+};
+
+const iconBuy = Icons.PlusMedium;
+const iconSell = Icons.MinusMedium;
+const iconSwap = Icons.BuyCryptoMedium;
+const iconReceive = Icons.ArrowBottomMedium;
+const iconSend = Icons.ArrowTopMedium;
+const iconAddAccount = Icons.WalletMedium;
+
+export default function useAssetActions({
+  currency,
+  accounts,
+}: useAssetActionsProps): {
+  mainActions: ActionButton[];
+} {
+  const { t } = useTranslation();
+  const readOnlyModeEnabled = useSelector(readOnlyModeEnabledSelector);
+  const hasAccounts = accounts?.length && accounts.length > 0;
+  const defaultAccount = useMemo(() => (accounts ? accounts[0] : undefined), [
+    accounts,
+  ]);
+
+  const swapSelectableCurrencies = useSelector(
+    swapSelectableCurrenciesSelector,
+  );
+  const availableOnSwap =
+    currency && swapSelectableCurrencies.includes(currency.id);
+
+  const rampCatalog = useRampCatalog();
+  const [canBeBought, canBeSold] = useMemo(() => {
+    if (!rampCatalog.value || !currency) {
+      return [false, false];
+    }
+
+    const onRampProviders = filterRampCatalogEntries(rampCatalog.value.onRamp, {
+      tickers: [currency.ticker],
+    });
+    const offRampProviders = filterRampCatalogEntries(
+      rampCatalog.value.offRamp,
+      {
+        tickers: [currency.ticker],
+      },
+    );
+
+    return [onRampProviders.length > 0, offRampProviders.length > 0];
+  }, [rampCatalog.value, currency]);
+
+  const actions = useMemo<ActionButton[]>(
+    () => [
+      ...(canBeBought
+        ? [
+            {
+              event: "TransferExchange",
+              label: t("exchange.buy.tabTitle"),
+              Icon: iconBuy,
+              navigationParams: [
+                NavigatorName.Exchange,
+                {
+                  screen: ScreenName.ExchangeBuy,
+                  params: {
+                    defaultTicker:
+                      currency &&
+                      currency.ticker &&
+                      currency.ticker.toUpperCase(),
+                  },
+                },
+              ],
+            },
+          ]
+        : []),
+      ...(canBeSold
+        ? [
+            {
+              event: "TransferExchange",
+              label: t("exchange.sell.tabTitle"),
+              Icon: iconSell,
+              navigationParams: [
+                NavigatorName.Exchange,
+                {
+                  screen: ScreenName.ExchangeSell,
+                  params: {
+                    defaultTicker:
+                      currency &&
+                      currency.ticker &&
+                      currency.ticker.toUpperCase(),
+                  },
+                },
+              ],
+            },
+          ]
+        : []),
+      ...(hasAccounts && !readOnlyModeEnabled
+        ? [
+            ...(availableOnSwap
+              ? [
+                  {
+                    event: "TransferSwap",
+                    label: t("transfer.swap.title"),
+                    Icon: iconSwap,
+                    navigationParams: [
+                      NavigatorName.Swap,
+                      {
+                        screen: ScreenName.Swap,
+                        params: { currencyId: currency?.id, defaultAccount },
+                      },
+                    ],
+                    disabled: defaultAccount?.balance.lte(0),
+                  },
+                ]
+              : []),
+            {
+              event: "TransferReceive",
+              label: t("transfer.receive.title"),
+              Icon: iconReceive,
+              navigationParams: [
+                NavigatorName.ReceiveFunds,
+                {
+                  screen: ScreenName.ReceiveSelectAccount,
+                  params: { selectedCurrency: currency },
+                },
+              ],
+            },
+            {
+              event: "TransferSend",
+              label: t("transfer.send.title"),
+              Icon: iconSend,
+              navigationParams: [
+                NavigatorName.SendFunds,
+                {
+                  screen: ScreenName.SendCoin,
+                  params: { selectedCurrency: currency },
+                },
+              ],
+              disabled: defaultAccount?.balance.lte(0),
+            },
+          ]
+        : [
+            ...(!readOnlyModeEnabled
+              ? [
+                  {
+                    event: "TransferAddAccount",
+                    label: t("addAccountsModal.ctaAdd"),
+                    Icon: iconAddAccount,
+                    navigationParams: [
+                      NavigatorName.AddAccounts,
+                      {
+                        screen: ScreenName.AddAccountsSelectCrypto,
+                        params: {
+                          filterCurrencyIds: currency
+                            ? [currency.id]
+                            : undefined,
+                        },
+                      },
+                    ],
+                  },
+                ]
+              : []),
+          ]),
+    ],
+    [
+      availableOnSwap,
+      canBeBought,
+      canBeSold,
+      currency,
+      defaultAccount,
+      hasAccounts,
+      readOnlyModeEnabled,
+      t,
+    ],
+  );
+
+  return {
+    mainActions: actions,
+  };
+}
