+++ apps/ledger-live-mobile/src/components/FabActions/index.tsx
@@ -12,23 +12,20 @@ import { useNavigation } from "@react-navigation/native";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 
-import { useRampCatalog } from "@ledgerhq/live-common/platform/providers/RampCatalogProvider/index";
-import { filterRampCatalogEntries } from "@ledgerhq/live-common/platform/providers/RampCatalogProvider/helpers";
 import { AccountLike, Account } from "@ledgerhq/types-live";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 
-
 import { Icons, QuickActionList } from "@ledgerhq/native-ui";
 
-import {
-  readOnlyModeEnabledSelector,
-  swapSelectableCurrenciesSelector,
-} from "../../reducers/settings";
+import { readOnlyModeEnabledSelector } from "../../reducers/settings";
 import { accountsCountSelector } from "../../reducers/accounts";
 import { NavigatorName, ScreenName } from "../../const";
 import FabAccountButtonBar from "./FabAccountButtonBar";
 import useAccountActions from "../../screens/Account/hooks/useAccountActions";
 import { useTheme } from "styled-components/native";
+import { QuickActionButtonProps } from "@ledgerhq/native-ui/components/cta/QuickAction/QuickActionButton";
+import { Linking } from "react-native";
+import useAssetActions from "./hooks/useAssetActions";
 
 export type ModalOnDisabledClickComponentProps = {
   account?: AccountLike;
@@ -81,7 +78,6 @@ const iconSell = Icons.MinusMedium;
 const iconSwap = Icons.BuyCryptoMedium;
 const iconReceive = Icons.ArrowBottomMedium;
 const iconSend = Icons.ArrowTopMedium;
-const iconAddAccount = Icons.WalletMedium;
 
 export const FabAccountMainActionsComponent: React.FC<FabAccountActionsProps> = ({
   account,
@@ -198,163 +194,62 @@ const FabMarketActionsComponent: React.FC<Props> = ({
   accounts,
   ...props
 }) => {
-  const { t } = useTranslation();
-  const readOnlyModeEnabled = useSelector(readOnlyModeEnabledSelector);
-  const hasAccounts = accounts?.length && accounts.length > 0;
-  const defaultAccount = useMemo(() => (accounts ? accounts[0] : undefined), [
-    accounts,
-  ]);
+  const { mainActions } = useAssetActions({ currency, accounts });
 
-  const swapSelectableCurrencies = useSelector(
-    swapSelectableCurrenciesSelector,
-  );
-  const availableOnSwap =
-    currency && swapSelectableCurrencies.includes(currency.id);
+  return <FabAccountButtonBar {...props} buttons={mainActions} />;
+};
 
-  const rampCatalog = useRampCatalog();
-  const [canBeBought, canBeSold] = useMemo(() => {
-    if (!rampCatalog.value || !currency) {
-      return [false, false];
-    }
+const FabAssetActionsComponent: React.FC<Props> = ({
+  currency,
+  accounts,
+  ...props
+}) => {
+  const navigation = useNavigation();
 
-    const onRampProviders = filterRampCatalogEntries(rampCatalog.value.onRamp, {
-      tickers: [currency.ticker],
-    });
-    const offRampProviders = filterRampCatalogEntries(
-      rampCatalog.value.offRamp,
-      {
-        tickers: [currency.ticker],
-      },
-    );
+  const { mainActions } = useAssetActions({ currency, accounts });
 
-    return [onRampProviders.length > 0, offRampProviders.length > 0];
-  }, [rampCatalog.value, currency]);
-
-  const actions = useMemo<ActionButton[]>(
-    () => [
-      ...(canBeBought
-        ? [
-            {
-              event: "TransferExchange",
-              label: t("exchange.buy.tabTitle"),
-              Icon: iconBuy,
-              navigationParams: [
-                NavigatorName.Exchange,
-                {
-                  screen: ScreenName.ExchangeBuy,
-                  params: {
-                    defaultTicker:
-                      currency &&
-                      currency.ticker &&
-                      currency.ticker.toUpperCase(),
-                  },
-                },
-              ],
-            },
-          ]
-        : []),
-      ...(canBeSold
-        ? [
-            {
-              event: "TransferExchange",
-              label: t("exchange.sell.tabTitle"),
-              Icon: iconSell,
-              navigationParams: [
-                NavigatorName.Exchange,
-                {
-                  screen: ScreenName.ExchangeSell,
-                  params: {
-                    defaultTicker:
-                      currency &&
-                      currency.ticker &&
-                      currency.ticker.toUpperCase(),
-                  },
-                },
-              ],
-            },
-          ]
-        : []),
-      ...(hasAccounts && !readOnlyModeEnabled
-        ? [
-            ...(availableOnSwap
-              ? [
-                  {
-                    event: "TransferSwap",
-                    label: t("transfer.swap.title"),
-                    Icon: iconSwap,
-                    navigationParams: [
-                      NavigatorName.Swap,
-                      {
-                        screen: ScreenName.Swap,
-                        params: { currencyId: currency?.id, defaultAccount },
-                      },
-                    ],
-                    disabled: defaultAccount?.balance.lte(0),
-                  },
-                ]
-              : []),
-            {
-              event: "TransferReceive",
-              label: t("transfer.receive.title"),
-              Icon: iconReceive,
-              navigationParams: [
-                NavigatorName.ReceiveFunds,
-                {
-                  screen: ScreenName.ReceiveSelectAccount,
-                  params: { selectedCurrency: currency },
-                },
-              ],
-            },
-            {
-              event: "TransferSend",
-              label: t("transfer.send.title"),
-              Icon: iconSend,
-              navigationParams: [
-                NavigatorName.SendFunds,
-                {
-                  screen: ScreenName.SendCoin,
-                  params: { selectedCurrency: currency },
-                },
-              ],
-              disabled: defaultAccount?.balance.lte(0),
-            },
-          ]
-        : [
-            ...(!readOnlyModeEnabled
-              ? [
-                  {
-                    event: "TransferAddAccount",
-                    label: t("addAccountsModal.ctaAdd"),
-                    Icon: iconAddAccount,
-                    navigationParams: [
-                      NavigatorName.AddAccounts,
-                      {
-                        screen: ScreenName.AddAccountsSelectCrypto,
-                        params: {
-                          filterCurrencyIds: currency
-                            ? [currency.id]
-                            : undefined,
-                        },
-                      },
-                    ],
-                  },
-                ]
-              : []),
-          ]),
-    ],
-    [
-      availableOnSwap,
-      canBeBought,
-      canBeSold,
-      currency,
-      defaultAccount,
-      hasAccounts,
-      readOnlyModeEnabled,
-      t,
-    ],
+  const onNavigate = useCallback(
+    (name: string, options?: any) => {
+      navigation.navigate(name, {
+        ...options,
+        params: {
+          ...(options ? options.params : {}),
+        },
+      });
+    },
+    [navigation],
   );
 
-  return <FabAccountButtonBar {...props} buttons={actions} />;
+  const onPress = useCallback(
+    (data: ActionButton) => {
+      const { navigationParams, linkUrl } = data;
+      if (linkUrl) {
+        Linking.openURL(linkUrl);
+      } else if (navigationParams) {
+        onNavigate(...navigationParams);
+      }
+    },
+    [onNavigate],
+  );
+
+  const quickActions: QuickActionButtonProps[] = mainActions.map(action => {
+    return {
+      Icon: action.Icon,
+      children: action.label,
+      onPress: () => onPress(action),
+      disabled: action.disabled,
+    };
+  });
+
+  return (
+    <QuickActionList
+      data={quickActions}
+      // Use two columns only when we have only two or four items, otherwise three columns
+      numColumns={
+        quickActions.length === 2 || quickActions.length === 4 ? 2 : 3
+      }
+    ></QuickActionList>
+  );
 };
 
 type FabActionsProps = {
@@ -449,5 +344,7 @@ const FabActions: React.FC<FabActionsProps> = ({
 export const FabAccountActions = memo(FabAccountActionsComponent);
 
 export const FabMarketActions = memo(FabMarketActionsComponent);
+
+export const FabAssetActions = memo(FabAssetActionsComponent);
 
 export default memo(FabActions);
